@@ -17,8 +17,10 @@
         }
     ],
     "require": {
-        "php": "^7.0",
-        "binsoul/net-mqtt": "^0.5",
+        "php": "^7.2",
+        "binsoul/net-mqtt": "^0.6",
+        "evenement/evenement": "^3",
+        "react/event-loop": "^1.1",
         "react/promise": "^2.7",
         "react/socket": "^1.3"
     },
